# openapi/aurum.yaml
openapi: 3.0.3
info:
  title: Aurum API
  version: 0.1.0
  description: API for retrieving market intelligence data, managing scenarios, and valuing PPAs.
servers:
  - url: https://api.aurum.local
    description: Local development gateway
  - url: https://api.aurum.prod
    description: Production
security:
  - oauth2: [aurum]
paths:
...
  /v1/metadata/dimensions:
    get:
      summary: List distinct dimension values useful for filters.
      operationId: listDimensions
      security:
        - oauth2: [aurum]
      parameters:
        - $ref: '#/components/parameters/AsOfDate'
        - $ref: '#/components/parameters/AssetClass'
        - $ref: '#/components/parameters/Iso'
        - $ref: '#/components/parameters/Location'
        - $ref: '#/components/parameters/Market'
        - $ref: '#/components/parameters/Product'
        - $ref: '#/components/parameters/Block'
        - $ref: '#/components/parameters/TenorType'
        - name: per_dim_limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 5000
        - name: prefix
          in: query
          schema:
            type: string
          description: Optional case-insensitive startswith filter applied to each dimension list
        - name: include_counts
          in: query
          schema:
            type: boolean
          description: Include per-dimension counts when true
      responses:
        '200':
          description: Dimension values.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DimensionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/metadata/locations:
    get:
      summary: List ISO locations
      operationId: listIsoLocations
      tags:
        - Metadata
      security:
        - oauth2: [aurum]
      parameters:
        - name: iso
          in: query
          required: false
          schema:
            type: string
          description: Filter by ISO code (e.g., PJM, CAISO)
        - name: prefix
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive startswith filter applied to id or name
      responses:
        '200':
          description: ISO location metadata returned.
          headers:
            ETag:
              description: Strong entity tag for cache validation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsoLocationsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/metadata/locations/{iso}/{location_id}:
    get:
      summary: Get ISO location
      operationId: getIsoLocation
      tags:
        - Metadata
      security:
        - oauth2: [aurum]
      parameters:
        - name: iso
          in: path
          required: true
          schema:
            type: string
        - name: location_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ISO location metadata returned.
          headers:
            ETag:
              description: Strong entity tag for cache validation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsoLocationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Location not found
  /v1/metadata/units:
    get:
      summary: List canonical currencies and units
      operationId: listUnits
      tags:
        - Metadata
        - Units
      security:
        - oauth2: [aurum]
      parameters:
        - name: prefix
          in: query
          required: false
          schema:
            type: string
          description: Optional case-insensitive startswith filter
      responses:
        '200':
          description: Canonical currency and unit lists returned.
          headers:
            ETag:
              description: Strong entity tag for cache validation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitsCanonicalResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/metadata/calendars:
    get:
      summary: List calendars
      operationId: listCalendars
      tags:
        - Metadata
      security:
        - oauth2: [aurum]
      responses:
        '200':
          description: Calendar metadata returned.
          headers:
            ETag:
              description: Strong entity tag for cache validation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/metadata/calendars/{name}/blocks:
    get:
      summary: List calendar blocks
      operationId: listCalendarBlocks
      tags:
        - Metadata
      security:
        - oauth2: [aurum]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Calendar block names returned.
          headers:
            ETag:
              description: Strong entity tag for cache validation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarBlocksResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Calendar not found
  /v1/metadata/calendars/{name}/hours:
    get:
      summary: Block hours for a date
      operationId: listCalendarHours
      tags:
        - Metadata
      security:
        - oauth2: [aurum]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: block
          in: query
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Calendar block hours returned.
          headers:
            ETag:
              description: Strong entity tag for cache validation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarHoursResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/metadata/calendars/{name}/expand:
    get:
      summary: Expand block over date range
      operationId: expandCalendarBlock
      tags:
        - Metadata
      security:
        - oauth2: [aurum]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: block
          in: query
          required: true
          schema:
            type: string
        - name: start
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Calendar block occurrences returned.
          headers:
            ETag:
              description: Strong entity tag for cache validation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarHoursResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.aurum.local/oauth2/auth
          tokenUrl: https://auth.aurum.local/oauth2/token
          scopes:
            aurum: Full access to Aurum APIs
  parameters:
    AsOfDate:
      name: asof
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Filter by as-of date.
    AssetClass:
      name: asset_class
      in: query
      schema:
        type: string
      required: false
    Iso:
      name: iso
      in: query
      schema:
        type: string
      required: false
    Location:
      name: location
      in: query
      schema:
        type: string
      required: false
    Market:
      name: market
      in: query
      schema:
        type: string
      required: false
    Product:
      name: product
      in: query
      schema:
        type: string
      required: false
    Block:
      name: block
      in: query
      schema:
        type: string
      required: false
    TenorType:
      name: tenor_type
      in: query
      schema:
        type: string
        enum: [MONTHLY, CALENDAR, SEASON, QUARTER]
      required: false
...
  responses:
    Unauthorized:
      description: Authentication failed or missing token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Meta:
      type: object
      properties:
        request_id:
          type: string
        query_time_ms:
          type: integer
        next_cursor:
          type: string
          nullable: true
        prev_cursor:
          type: string
          nullable: true
...
    DimensionsData:
      type: object
      properties:
        asset_class:
          type: array
          items:
            type: string
        iso:
          type: array
          items:
            type: string
        location:
          type: array
          items:
            type: string
        market:
          type: array
          items:
            type: string
        product:
          type: array
          items:
            type: string
        block:
          type: array
          items:
            type: string
        tenor_type:
          type: array
          items:
            type: string
    DimensionsResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          $ref: '#/components/schemas/DimensionsData'
        counts:
          $ref: '#/components/schemas/DimensionsCountData'
          nullable: true
    DimensionCount:
      type: object
      required: [value, count]
      properties:
        value:
          type: string
        count:
          type: integer
    DimensionsCountData:
      type: object
      properties:
        asset_class:
          type: array
          items:
            $ref: '#/components/schemas/DimensionCount'
        iso:
          type: array
          items:
            $ref: '#/components/schemas/DimensionCount'
        location:
          type: array
          items:
            $ref: '#/components/schemas/DimensionCount'
        market:
          type: array
          items:
            $ref: '#/components/schemas/DimensionCount'
        product:
          type: array
          items:
            $ref: '#/components/schemas/DimensionCount'
        block:
          type: array
          items:
            $ref: '#/components/schemas/DimensionCount'
        tenor_type:
          type: array
          items:
            $ref: '#/components/schemas/DimensionCount'
    IsoLocationOut:
      type: object
      required: [iso, location_id]
      properties:
        iso:
          type: string
        location_id:
          type: string
        location_name:
          type: string
          nullable: true
        location_type:
          type: string
          nullable: true
        zone:
          type: string
          nullable: true
        hub:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
    IsoLocationsResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: array
          items:
            $ref: '#/components/schemas/IsoLocationOut'
    IsoLocationResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          $ref: '#/components/schemas/IsoLocationOut'
    UnitsCanonical:
      type: object
      required: [currencies, units]
      properties:
        currencies:
          type: array
          items:
            type: string
        units:
          type: array
          items:
            type: string
    UnitsCanonicalResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          $ref: '#/components/schemas/UnitsCanonical'
...
    CalendarOut:
      type: object
      required: [name, timezone]
      properties:
        name:
          type: string
        timezone:
          type: string
    CalendarsResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: array
          items:
            $ref: '#/components/schemas/CalendarOut'
    CalendarBlocksResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: array
          items:
            type: string
    CalendarHoursResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: array
          items:
            type: string
            format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        meta:
          $ref: '#/components/schemas/Meta'